[#64] feat: 파일 경로 수정 및 충돌 해결
