+++ src/app/signin/page.tsx
@@ -1,9 +1,16 @@
-'use client';
+import authOptions from '@/app/api/auth/[...nextauth]/options';
+import SocialSigninButton from '@/components/atoms/Button/SocialSigninButton';
+import { getServerSession } from 'next-auth';
+import { getProviders } from 'next-auth/react';
 
-import RegisterFormProvider from '@/providers/FormProvider/RegisterFormProvider';
+async function Signin() {
+  const session = await getServerSession(authOptions);
+  const providers = await getProviders();
+  if (session) {
+    return { redirect: { destination: '/' } };
+  }
 
-function Signin() {
-  return <RegisterFormProvider></RegisterFormProvider>;
+  return <SocialSigninButton providers={providers} />;
 }
 
 export default Signin;
